Moving HyperparameterRanges to searcher/utils/
