+++ examples/launch_standalone_bayesian_optimization.py
@@ -17,7 +17,7 @@ from syne_tune.config_space import uniform, randint, choice
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
     dictionarize_objective,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_factory import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
     make_hyperparameter_ranges,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.test_objects import (
