+++ syne_tune/optimizer/schedulers/neuralbands/neuralband_supplement.py
@@ -21,7 +21,7 @@ from syne_tune.backend.trial_status import Trial
 from syne_tune.optimizer.schedulers.searchers.searcher import BaseSearcher
 from syne_tune.optimizer.schedulers.searchers.searcher_factory import searcher_factory
 from syne_tune.config_space import cast_config_values
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_factory import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
     make_hyperparameter_ranges,
 )
 from syne_tune.optimizer.schedulers.neuralbands.networks import Exploitation
