+++ syne_tune/optimizer/schedulers/searchers/bayesopt/datatypes/config_ext.py
@@ -14,7 +14,7 @@ from typing import Tuple
 import copy
 
 from syne_tune.config_space import randint
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
