+++ syne_tune/optimizer/schedulers/searchers/bayesopt/datatypes/tuning_job_state.py
@@ -19,7 +19,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
     MetricValues,
     INTERNAL_METRIC_NAME,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
