+++ syne_tune/optimizer/schedulers/searchers/bayesopt/gpautograd/independent/posterior_state.py
@@ -19,7 +19,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.gpautograd.posterior_stat
     PosteriorStateWithSampleJoint,
     GaussProcPosteriorState,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_impl import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_impl import (
     decode_extended_features,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.gpautograd.kernel import (
