+++ syne_tune/optimizer/schedulers/searchers/bayesopt/tuning_algorithms/base_classes.py
@@ -28,7 +28,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
     Configuration,
     INTERNAL_METRIC_NAME,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.tuning_job_state import (
