+++ syne_tune/optimizer/schedulers/searchers/utils/hp_ranges_factory.py
@@ -13,10 +13,10 @@
 from typing import Dict, Optional, List
 import logging
 
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_impl import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_impl import (
     HyperparameterRangesImpl,
 )
