+++ syne_tune/optimizer/schedulers/searchers/utils/hp_ranges_impl.py
@@ -19,7 +19,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
     Hyperparameter,
     Configuration,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.scaling import (
