+++ tst/schedulers/bayesopt/test_bo_algorithm_components.py
@@ -12,7 +12,7 @@
 # permissions and limitations under the License.
 import pytest
 
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_factory import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
     make_hyperparameter_ranges,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
