+++ tst/schedulers/bayesopt/test_bo_algorithm_functions.py
@@ -20,7 +20,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algo
 from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.duplicate_detector import (
     DuplicateDetectorIdentical,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_factory import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
     make_hyperparameter_ranges,
 )
 from syne_tune.config_space import uniform, randint, choice
