+++ tst/schedulers/bayesopt/test_duplicate_detector.py
@@ -16,7 +16,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.duplicate_detector 
     DuplicateDetectorIdentical,
     DuplicateDetectorNoDetection,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.hp_ranges_factory import (
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
     make_hyperparameter_ranges,
 )
 from syne_tune.config_space import uniform, randint, choice
